[Impeller] Reland: Refactor color source resolution to use explicit factory types (#37656)" (#37673)
This reverts commit d721647766ddf697d26f05f4ac56d4142ef77742.
