+++ impeller/aiks/color_source_factory.cc
@@ -1,0 +1,11 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include "impeller/aiks/color_source_factory.h"
+
+namespace impeller {
+
+ColorSourceFactory::~ColorSourceFactory() = default;
+
+}  // namespace impeller
